requirements.txt
# AutoGen 0.4 核心依赖
autogen-agentchat>=0.4.0
autogen-ext>=0.4.0

# LLM 提供商
openai>=1.0.0
anthropic>=0.8.0

# 数据验证和处理
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 异步 HTTP 客户端
aiohttp>=3.9.0
httpx>=0.25.0

# 日志和工具
colorlog>=6.8.0
rich>=13.7.0
